Rename CommPkg folder to scope
