+++ src/pages/Scope/DetailsCard.test.tsx
