+++ src/pages/Scope/DetailsCard.tsx
