+++ src/pages/Scope/DetailsCardShell.tsx
