+++ src/pages/Scope/Scope/Scope.tsx
